refactor(party): split repository.go by domain
Split the 966-line party repository into domain-specific files:
- repository.go: struct, constructor, shared helpers, core party CRUD, ListParties
- repository_association.go: association operations (save, find, update, circular check, participants)
- repository_demographic.go: demographic save/find
- repository_reference.go: reference save/find
- repository_bank_relation.go: bank relation save/find

Removes party repository from knownOversizedFiles (now 616 lines, under 800 limit).
