+++ services/party/adapters/persistence/repository.go
@@ -1,4 +1,3 @@
-//meridian:large-file - known oversized file; split tracked in backlog
 package persistence
 
 import (
@@ -6,7 +5,6 @@ import (
 	"encoding/base64"
 	"encoding/json"
 	"errors"
-	"fmt"
 	"strings"
 	"time"
 
@@ -248,7 +246,7 @@ func (r *Repository) Save(ctx context.Context, party *domain.Party) error {
 }
 
 // SaveInTx creates or updates a party within an existing database transaction.
-// Use this when the caller manages the transaction boundary — for example, to include
+// Use this when the caller manages the transaction boundary - for example, to include
 // an outbox event write in the same transaction for atomic event delivery.
 func (r *Repository) SaveInTx(ctx context.Context, party *domain.Party, tx *gorm.DB) error {
 	entity := toEntity(ctx, party)
@@ -531,321 +529,6 @@ func isDuplicateKeyError(err error) bool {
 		strings.Contains(errStr, "unique constraint")
 }
 
-// AssociationInput holds optional fields for creating a party association.
-type AssociationInput struct {
-	Metadata      *string
-	Status        string
-	EffectiveFrom *time.Time
-	EffectiveTo   *time.Time
-}
-
-// SaveAssociation saves a party association.
-// Returns ErrAssociationExists if an association already exists between the parties.
-func (r *Repository) SaveAssociation(ctx context.Context, partyID, relatedPartyID uuid.UUID, relationshipType string) (uuid.UUID, error) {
-	return r.SaveAssociationWithInput(ctx, partyID, relatedPartyID, relationshipType, nil)
-}
-
-// SaveAssociationWithInput saves a party association with optional metadata and lifecycle fields.
-// Returns ErrAssociationExists if an association already exists between the parties.
-func (r *Repository) SaveAssociationWithInput(ctx context.Context, partyID, relatedPartyID uuid.UUID, relationshipType string, input *AssociationInput) (uuid.UUID, error) {
-	now := time.Now()
-	associationID := uuid.New()
-	entity := &PartyAssociationEntity{
-		ID:               associationID,
-		PartyID:          partyID,
-		RelatedPartyID:   relatedPartyID,
-		RelationshipType: relationshipType,
-		Status:           "ACTIVE",
-		EffectiveFrom:    now,
-		CreatedAt:        now,
-		UpdatedAt:        now,
-	}
-
-	if input != nil {
-		if input.Metadata != nil {
-			entity.Metadata = input.Metadata
-		}
-		if input.Status != "" {
-			entity.Status = input.Status
-		}
-		if input.EffectiveFrom != nil {
-			entity.EffectiveFrom = *input.EffectiveFrom
-		}
-		entity.EffectiveTo = input.EffectiveTo
-	}
-
-	err := r.withTenantTransaction(ctx, func(tx *gorm.DB) error {
-		return tx.Create(entity).Error
-	})
-	if err != nil {
-		if isDuplicateKeyError(err) {
-			return uuid.Nil, ErrAssociationExists
-		}
-		return uuid.Nil, err
-	}
-	return associationID, nil
-}
-
-// FindAssociations retrieves all associations for a party
-func (r *Repository) FindAssociations(ctx context.Context, partyID uuid.UUID) ([]PartyAssociationEntity, error) {
-	var associations []PartyAssociationEntity
-	err := r.withTenantTransaction(ctx, func(tx *gorm.DB) error {
-		return tx.Where("party_id = ?", partyID).Find(&associations).Error
-	})
-	return associations, err
-}
-
-// UpdateAssociation updates an association's relationship type and returns the updated entity.
-// Returns an error if the association doesn't exist.
-func (r *Repository) UpdateAssociation(ctx context.Context, associationID uuid.UUID, relationshipType string) (*PartyAssociationEntity, error) {
-	var entity PartyAssociationEntity
-	err := r.withTenantTransaction(ctx, func(tx *gorm.DB) error {
-		result := tx.Model(&PartyAssociationEntity{}).
-			Where("id = ?", associationID).
-			Updates(map[string]interface{}{
-				"relationship_type": relationshipType,
-				"updated_at":        time.Now(),
-			})
-		if result.Error != nil {
-			return result.Error
-		}
-		if result.RowsAffected == 0 {
-			return gorm.ErrRecordNotFound
-		}
-		// Load the updated entity to return party_id and related_party_id
-		return tx.Where("id = ?", associationID).First(&entity).Error
-	})
-	if err != nil {
-		return nil, err
-	}
-	return &entity, nil
-}
-
-// CheckCircularAssociation checks if adding this association would create a circular reference
-func (r *Repository) CheckCircularAssociation(ctx context.Context, partyID, relatedPartyID uuid.UUID) (bool, error) {
-	// Simple check: verify they're not the same and no direct reverse relationship exists
-	if partyID == relatedPartyID {
-		return true, nil
-	}
-
-	var count int64
-	err := r.withTenantTransaction(ctx, func(tx *gorm.DB) error {
-		// Check if reverse relationship exists
-		return tx.Model(&PartyAssociationEntity{}).
-			Where("party_id = ? AND related_party_id = ?", relatedPartyID, partyID).
-			Count(&count).Error
-	})
-	if err != nil {
-		return false, err
-	}
-	return count > 0, nil
-}
-
-// SaveDemographic saves or updates demographic data for a party
-func (r *Repository) SaveDemographic(ctx context.Context, partyID uuid.UUID, socioEconomicData, employmentHistory string) error {
-	return r.withTenantTransaction(ctx, func(tx *gorm.DB) error {
-		// Check if exists
-		var existing PartyDemographicEntity
-		result := tx.Where("party_id = ?", partyID).First(&existing)
-
-		// Prepare strings for JSONB columns
-		// If already valid JSON, store as-is; otherwise wrap as JSON string
-		socioEconStr := toJSONB(socioEconomicData)
-		empHistoryStr := toJSONB(employmentHistory)
-		socioEcon := &socioEconStr
-		empHistory := &empHistoryStr
-
-		if errors.Is(result.Error, gorm.ErrRecordNotFound) {
-			// Create new
-			entity := &PartyDemographicEntity{
-				ID:                uuid.New(),
-				PartyID:           partyID,
-				SocioEconomicData: socioEcon,
-				EmploymentHistory: empHistory,
-				UpdatedAt:         time.Now(),
-			}
-			return tx.Create(entity).Error
-		}
-
-		// Update existing
-		return tx.Model(&PartyDemographicEntity{}).
-			Where("party_id = ?", partyID).
-			Updates(map[string]interface{}{
-				"socio_economic_data": socioEcon,
-				"employment_history":  empHistory,
-				"updated_at":          time.Now(),
-			}).Error
-	})
-}
-
-// FindDemographic retrieves demographic data for a party.
-// Returns (nil, nil) if no demographic data exists for the party.
-func (r *Repository) FindDemographic(ctx context.Context, partyID uuid.UUID) (*PartyDemographicEntity, error) {
-	var demographic PartyDemographicEntity
-	var found bool
-	err := r.withTenantTransaction(ctx, func(tx *gorm.DB) error {
-		result := tx.Where("party_id = ?", partyID).First(&demographic)
-		if errors.Is(result.Error, gorm.ErrRecordNotFound) {
-			found = false
-			return nil // Not an error, just no demographic data
-		}
-		if result.Error != nil {
-			return result.Error
-		}
-		found = true
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-	if !found {
-		return nil, nil //nolint:nilnil // nil,nil signals "not found" without error
-	}
-	return &demographic, nil
-}
-
-// ReferenceInput represents input for saving a reference.
-type ReferenceInput struct {
-	RefType          string
-	RefValue         string
-	IssuingAuthority string
-	ExpiryDate       string
-}
-
-// SaveReference saves party reference data
-func (r *Repository) SaveReference(ctx context.Context, partyID uuid.UUID, refType, refValue, issuingAuthority, expiryDate string) error {
-	return r.SaveReferences(ctx, partyID, []ReferenceInput{{
-		RefType:          refType,
-		RefValue:         refValue,
-		IssuingAuthority: issuingAuthority,
-		ExpiryDate:       expiryDate,
-	}})
-}
-
-// SaveReferences saves multiple party references in a single transaction.
-func (r *Repository) SaveReferences(ctx context.Context, partyID uuid.UUID, refs []ReferenceInput) error {
-	if len(refs) == 0 {
-		return nil
-	}
-
-	return r.withTenantTransaction(ctx, func(tx *gorm.DB) error {
-		for _, ref := range refs {
-			entity := &PartyReferenceEntity{
-				ID:             uuid.New(),
-				PartyID:        partyID,
-				ReferenceType:  ref.RefType,
-				ReferenceValue: ref.RefValue,
-				CreatedAt:      time.Now(),
-			}
-
-			if ref.IssuingAuthority != "" {
-				entity.IssuingAuthority = &ref.IssuingAuthority
-			}
-			if ref.ExpiryDate != "" {
-				parsedDate, err := time.Parse("2006-01-02", ref.ExpiryDate)
-				if err == nil {
-					entity.ExpiryDate = &parsedDate
-				}
-			}
-
-			if err := tx.Create(entity).Error; err != nil {
-				return err
-			}
-		}
-		return nil
-	})
-}
-
-// FindReferences retrieves all references for a party
-func (r *Repository) FindReferences(ctx context.Context, partyID uuid.UUID) ([]PartyReferenceEntity, error) {
-	var references []PartyReferenceEntity
-	err := r.withTenantTransaction(ctx, func(tx *gorm.DB) error {
-		return tx.Where("party_id = ?", partyID).Find(&references).Error
-	})
-	return references, err
-}
-
-// SaveBankRelation saves or updates bank relationship data
-func (r *Repository) SaveBankRelation(ctx context.Context, partyID uuid.UUID, accountOfficerID, relationshipManagerID, assignedBranch string) error {
-	return r.withTenantTransaction(ctx, func(tx *gorm.DB) error {
-		// Check if exists
-		var existing PartyBankRelationEntity
-		result := tx.Where("party_id = ?", partyID).First(&existing)
-
-		var aoID, rmID, branch *string
-		if accountOfficerID != "" {
-			aoID = &accountOfficerID
-		}
-		if relationshipManagerID != "" {
-			rmID = &relationshipManagerID
-		}
-		if assignedBranch != "" {
-			branch = &assignedBranch
-		}
-
-		if errors.Is(result.Error, gorm.ErrRecordNotFound) {
-			// Create new
-			entity := &PartyBankRelationEntity{
-				ID:                    uuid.New(),
-				PartyID:               partyID,
-				AccountOfficerID:      aoID,
-				RelationshipManagerID: rmID,
-				AssignedBranch:        branch,
-				UpdatedAt:             time.Now(),
-			}
-			return tx.Create(entity).Error
-		}
-
-		// Update existing
-		return tx.Model(&PartyBankRelationEntity{}).
-			Where("party_id = ?", partyID).
-			Updates(map[string]interface{}{
-				"account_officer_id":      aoID,
-				"relationship_manager_id": rmID,
-				"assigned_branch":         branch,
-				"updated_at":              time.Now(),
-			}).Error
-	})
-}
-
-// FindBankRelation retrieves bank relationship data for a party.
-// Returns (nil, nil) if no bank relation data exists for the party.
-func (r *Repository) FindBankRelation(ctx context.Context, partyID uuid.UUID) (*PartyBankRelationEntity, error) {
-	var bankRelation PartyBankRelationEntity
-	var found bool
-	err := r.withTenantTransaction(ctx, func(tx *gorm.DB) error {
-		result := tx.Where("party_id = ?", partyID).First(&bankRelation)
-		if errors.Is(result.Error, gorm.ErrRecordNotFound) {
-			found = false
-			return nil // Not an error, just no bank relation data
-		}
-		if result.Error != nil {
-			return result.Error
-		}
-		found = true
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-	if !found {
-		return nil, nil //nolint:nilnil // nil,nil signals "not found" without error
-	}
-	return &bankRelation, nil
-}
-
-// ListParticipants retrieves all ACTIVE associations where the given orgPartyID is the related_party_id
-// (i.e., the org is the syndicate host and participants point to it) with the given relationship type.
-func (r *Repository) ListParticipants(ctx context.Context, orgPartyID uuid.UUID, relationshipType string) ([]PartyAssociationEntity, error) {
-	var associations []PartyAssociationEntity
-	err := r.withTenantTransaction(ctx, func(tx *gorm.DB) error {
-		return tx.Where("related_party_id = ? AND relationship_type = ? AND status = ?",
-			orgPartyID, relationshipType, "ACTIVE").
-			Find(&associations).Error
-	})
-	return associations, err
-}
-
 // ListParties retrieves a paginated list of parties with optional filtering.
 // Results are ordered by created_at DESC, id DESC (newest first) for stable pagination.
 // The context must contain the tenant ID for schema routing.
@@ -930,37 +613,4 @@ func (r *Repository) ListParties(ctx context.Context, params ListPartiesParams) 
 		return nil, err
 	}
 	return result, nil
-}
-
-// GetStructuringData retrieves the metadata JSONB for a specific association.
-// Returns an empty map (not an error) if no association is found.
-func (r *Repository) GetStructuringData(ctx context.Context, partyID, orgPartyID uuid.UUID, relationshipType string) (map[string]interface{}, error) {
-	var entity PartyAssociationEntity
-	var found bool
-	err := r.withTenantTransaction(ctx, func(tx *gorm.DB) error {
-		result := tx.Where("party_id = ? AND related_party_id = ? AND relationship_type = ?",
-			partyID, orgPartyID, relationshipType).
-			First(&entity)
-		if errors.Is(result.Error, gorm.ErrRecordNotFound) {
-			found = false
-			return nil
-		}
-		if result.Error != nil {
-			return result.Error
-		}
-		found = true
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-	if !found || entity.Metadata == nil {
-		return map[string]interface{}{}, nil
-	}
-
-	var metadata map[string]interface{}
-	if err := json.Unmarshal([]byte(*entity.Metadata), &metadata); err != nil {
-		return nil, fmt.Errorf("unmarshal association metadata: %w", err)
-	}
-	return metadata, nil
 }
